BCM59055 SAR ADC: Modification
[Defect / Enhancement / New Feature]
Defect

[Problem]
1. while setting RTM channel, channel number was not shifted properly
2. BCM59055_ADCCTRL1_RTM_ENABLE should be cleared in order to start RTM
3. return type of bcm59055_saradc_request_rtm() should be int
4. INIT_COMPLETION needed to be moved above interrupt enabling and/or setting RTM start bit
5. CLOCK_CYCLE_IN_USEC is better to set 32 as accurate value is 31.25uSec

[Solution]
Done the same.

[Reviewers]
Shameer P, Henrik J

[Testing]
    [Test Cases]

    [Label/CL# Synced To]

    [Platform]

[BROADCOM INTERNAL NOTES]

+++ include/linux/broadcom/bcm59055-adc.h
@@ -121,7 +121,7 @@ extern int bcm59055_saradc_stop_burst_mode(void);
 extern int bcm59055_saradc_enable_ntc_block(void);
 extern int bcm59055_saradc_disable_ntc_block(void);
 extern int bcm59055_saradc_read_data(int sel);
-extern u16 bcm59055_saradc_request_rtm(int ch_sel);
+extern int bcm59055_saradc_request_rtm(int ch_sel);
 extern int bcm59055_saradc_set_rtm_delay(int delay);
 #endif
